// types/agents.ts
export interface Agent {
  COMPID: string
  AGNTID: string
  AGNTDSC: string
  ADDRL1: string
  ADDRL2: string
  City: string
  ZIPCODE: string
  Phone: string
  Fax: string
  eMail: string
  WebSite: string
  FEDTXID: string
  STETXID: string
  CNTYCOD: string
  STAID: string
  CRTUSR: string
  CRTDAT: string
  CRTTIM: string
  CHGUSR: string
  CHGDAT: string
  CHGTIM: string
  RowNum: number
} 